Allow for storing the maximum vanilla survival XP
As per the wiki, the maximum amount of XP *levels* you can have in survival vanilla is 238,609,312, at which point the next level would take more than the integer limit of XP to get to. By storing the XP as integer, the XP Block can only store ~15000 levels of XP, instead of 238,609,312 levels. This change fixes that.